requirements.txt
Django
django-filter
requests
celery>=5.0
redis>=3.5.3
djangorestframework>=3.12
markdown>=3.3.3
PyJWT==2.1.0
gunicorn
